Move frame transformations to Tensor lib
