+++ src/DataStructures/Tensor/EagerMath/FrameTransform.cpp
@@ -1,7 +1,7 @@
 // Distributed under the MIT License.
 // See LICENSE.txt for details.
 
-#include "PointwiseFunctions/GeneralRelativity/Transform.hpp"
+#include "DataStructures/Tensor/EagerMath/FrameTransform.hpp"
 
 #include <limits>
