+++ src/PointwiseFunctions/GeneralRelativity/CMakeLists.txt
@@ -27,7 +27,6 @@ spectre_target_sources(
   SpatialMetric.cpp
   TimeDerivativeOfSpacetimeMetric.cpp
   TimeDerivativeOfSpatialMetric.cpp
-  Transform.cpp
   WeylElectric.cpp
   WeylMagnetic.cpp
   WeylPropagating.cpp
@@ -60,7 +59,6 @@ spectre_target_headers(
   TagsDeclarations.hpp
   TimeDerivativeOfSpacetimeMetric.hpp
   TimeDerivativeOfSpatialMetric.hpp
-  Transform.hpp
   WeylElectric.hpp
   WeylMagnetic.hpp
   WeylPropagating.hpp
