+++ tests/Unit/DataStructures/Tensor/EagerMath/CMakeLists.txt
@@ -8,6 +8,7 @@ set(LIBRARY_SOURCES
   Test_Determinant.cpp
   Test_DeterminantAndInverse.cpp
   Test_DotProduct.cpp
+  Test_FrameTransform.cpp
   Test_Magnitude.cpp
   Test_Norms.cpp
   Test_OrthonormalOneform.cpp
@@ -21,6 +22,7 @@ target_link_libraries(
   ${LIBRARY}
   PRIVATE
   DataStructures
+  DataStructuresHelpers
   DomainHelpers
   ErrorHandling
   GeneralRelativity
