+++ tests/Unit/PointwiseFunctions/GeneralRelativity/CMakeLists.txt
@@ -13,7 +13,6 @@ set(LIBRARY_SOURCES
   Test_Psi4.cpp
   Test_Ricci.cpp
   Test_Tags.cpp
-  Test_Transform.cpp
   Test_WeylElectric.cpp
   Test_WeylMagnetic.cpp
   Test_WeylPropagating.cpp
